add CartItemPolicy to restrict cart item deletion to owner also now blade page only shows logged in users items and not other users items.

+++ app/Http/Controllers/CartController.php
@@ -55,7 +55,8 @@ class CartController extends Controller
 
   public function delete(CartItems $cartItem)
   {
-    // dd($cartItem);
+
+    $this->authorize('delete',$cartItem);
     $cartItem->delete();
     return redirect()->route('cart.index');
   }
